Regenerate Makefile.in for zstd dependency
The version of Makefile.in that landed in https://crrev.com/c/4722191
had a Makefile.in that was generated from a previous version of
Makefile.am where the libzstd dependency wasn't optional. That's causing
some problems (see: https://crrev.com/c/5193965) and is a simple
mistake.

I generated this CL by simply running `automake`, so it simply makes the
checked in generated Makefile.in based on the checked in source
Makefile.am

Change-Id: Iabb4a99bfac3f5ef6067a140bd373c9fb894878a
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/5200626
Reviewed-by: Mike Frysinger <vapier@chromium.org>