Allows setting maximum number of threads to process.
By default, there is no limit and all threads will be processed. If set to a positive integer (via set_max_thread_count), the minidump processor will stop once that limit is reached. The limit will be ignored if the crashing/requesting thread is above the specified limit.

The original thread count is stored in the ProcessState and can be
acessed via the original_thread_count() method.

Change-Id: I6f0f173e5ad163aa7b2be76780b4a86e4932fc17
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/5641257
Reviewed-by: Joshua Peraza <jperaza@chromium.org>