fix some invalid uses of nullptr
Some Windows code was misusing NULL, so when it was changed to nullptr,
the extra compile-time checks caught it and broke.

Since crash_id_ is an integer, initialize it to 0, not a pointer.
  client/windows/crash_generation/client_info.cc(60,7): error:
  cannot initialize a member subobject of type 'DWORD' (aka 'unsigned long') with an rvalue of type 'std::nullptr_t'
   60 |       crash_id_(nullptr) {

Casting a pointer to the MINIDUMP_TYPE enum doesn't make sense,
especially when the enum with value 0 is a valid value.  Switch
to MiniDumpNormal since that's what NULL (0) was doing.
  client/windows/crash_generation/crash_generation_client.cc(209,23): error:
  static_cast from 'std::nullptr_t' to 'MINIDUMP_TYPE' (aka '_MINIDUMP_TYPE') is not allowed
  209 |                       static_cast<MINIDUMP_TYPE>(nullptr), nullptr, nullptr,

While these are ULONG_PTR, MSDN says they're status codes & byte counts,
so initialize both to 0 rather than a null pointer.
  client/windows/crash_generation/crash_generation_server.cc(568,26): error:
  assigning to 'ULONG_PTR' (aka 'unsigned long long') from incompatible type 'std::nullptr_t'
  568 |   overlapped_.Internal = nullptr;
  client/windows/crash_generation/crash_generation_server.cc(569,30): error:
  assigning to 'ULONG_PTR' (aka 'unsigned long long') from incompatible type 'std::nullptr_t'
  569 |   overlapped_.InternalHigh = nullptr;

Since handle_ is an integer, initialize it to 0.
  client/windows/crash_generation/minidump_generator.cc(109,7): error:
  cannot initialize a member subobject of type 'ULONG64' (aka 'unsigned long long') with an rvalue of type 'std::nullptr_t'
  109 |       handle_(nullptr) {

While Linux defines AppMemory.ptr as void*, Windows defines it as
ULONG64.  Cast nullptr to that to match other ptr code in this file.
  client/windows/handler/exception_handler.cc(256,28): error:
  assigning to 'ULONG64' (aka 'unsigned long long') from incompatible type 'std::nullptr_t'
  256 |   instruction_memory.ptr = nullptr;

The last arg to WinHttpConnect is a reserved integer, not a pointer,
so passing it NULL was incorrect -- change to 0 per MSDN.
  tools/windows/converter_exe/winhttp_client.cc(150,42): error:
  cannot initialize a parameter of type 'DWORD' (aka 'unsigned long') with an rvalue of type 'std::nullptr_t'
  150 |                                          nullptr));

We need to pass explicit 0 for DWORD_PTR types.
  tools/windows/converter_exe/wininet_client.cc(175,39): error:
  cannot initialize a parameter of type 'DWORD_PTR' (aka 'unsigned long long') with an rvalue of type 'std::nullptr_t'
  175 |                                       nullptr));

Change-Id: Id5f5b65a4279fe7aa1cbc1b99b18813d6a150117
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/6388180
Reviewed-by: Lei Zhang <thestig@chromium.org>