Strip static and return type from Windows symbols
DIA's demangler includes "static <return type>" in the demangled
(undecorated) signature for functions with internal linkage, which is
inconsistent with other platforms and other demangled function
signatures. It does this regardless of the flags passed to
`get_undecoratedNameEx()`.

This change makes `dump_syms` strip the start of symbols starting with
the "static" keyword, removing it and the return type and leaving only
the function name and arguments.

Using LLVM's demangler as a backup was considered as an alternative,
but llvm::microsoftDemangle()'s underlying logic in Demangler::parse()
requires a mangled (decorated) symbol starting with '?' [1]. Only public
symbols are mangled [2], and the symbols affected by the `static` issue
are always private (since they have internal linkage).

I tested this change locally by running dump_syms on chrome.exe.pdb
and comparing its output before and after this change. I've uploaded
the diff here (Google-internal link):
https://drive.google.com/file/d/1a3cH0wxeiJUZkRw0aAKH-vuNBQ3AG_V6/view?usp=sharing&resourcekey=0-hWswlsLwJtQv67nfW5g7Uw

[1] https://llvm.org/doxygen/MicrosoftDemangle_8cpp_source.html line 891
[2] https://learn.microsoft.com/en-us/windows-hardware/drivers/debugger/public-and-private-symbols

Bug: 417834385
Change-Id: I75d195d498e786696b727376fd96faffcae3af04
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/7226762
Reviewed-by: Ivan Penkov <ivanpe@google.com>
Reviewed-by: Mark Mentovai <mark@chromium.org>